refactor: move PinDocument and friends to repo::mem
